refactor: colocate CaptureScreen sub-components with the screen
Move CaptureInput, CaptureList, SessionForm from shared components/
into screens/CaptureScreen/ — they are only used by CaptureScreen.

src/screens/CaptureScreen/
├── CaptureScreen.tsx
├── CaptureInput.tsx
├── CaptureList.tsx
└── SessionForm.tsx

components/ now only holds shared primitives (BottomTabBar).
