+++ packages/web/src/screens/CaptureScreen/CaptureInput.tsx
